Import live.2006.02.15.tar.gz
- Changed "GetFileSize()" (defined in "liveMedia/include/InputFile.hh") and
  "ByteStreamFileSource::fileSize()" to return a "u_int64_t", to allow for
  (potentially) large files.
- Changed the "microseconds" parameter to "TaskScheduler::scheduleDelayedTask()"
  from "int" to "int64_t".  This allows for durations greater than the previous
  ~2000 second limit.
- Updated "RTSPClient" to support handling RTSP requests sent by the server.
  (Some servers send periodic requests back to the client as a keep-alive test.)
  At present, we just respond to such requests with a "not allowed" error.
  (Thanks to Khanh Mai for this change.)
- Updated "MultiFramedRTPSource::doStopGettingFrames()" to reset the
  "ReorderingPacketBuffer".  (Thanks to David Bertrand for this fix.)
- Corrected a small bug in the previous "RTSPClient" modification
  (noticed by Glen Gray).
- Fixed a bug in "ADTSAudioFileSource" when parsing the optional "crc_check"
  field.  (Thanks to Paul Li for noticing this.)

+++ BasicUsageEnvironment/BasicTaskScheduler0.cpp
@@ -52,11 +52,11 @@ BasicTaskScheduler0::~BasicTaskScheduler0() {
   delete fReadHandlers;
 }
 
-TaskToken BasicTaskScheduler0::scheduleDelayedTask(int microseconds,
+TaskToken BasicTaskScheduler0::scheduleDelayedTask(int64_t microseconds,
 						 TaskFunc* proc,
 						 void* clientData) {
   if (microseconds < 0) microseconds = 0;
-  DelayInterval timeToDelay(microseconds/1000000, microseconds%1000000);
+  DelayInterval timeToDelay((long)(microseconds/1000000), (long)(microseconds%1000000));
   AlarmHandler* alarmHandler = new AlarmHandler(proc, clientData, timeToDelay);
   fDelayQueue.addEntry(alarmHandler);
